Updated scaling of holdables
The parent Area2D of holdables is now a 1x1 scale, and the child CollisionShape2D and Sprite is now 0.8x0.8. This allows for the entire holdable to be flipped along the x axis without causing the sprite to enlarge from 0.8 to 1 as it did when the Area2D itself was set to a 0.8x0.8 scale.